# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: EveryReview API
  version: "1.0"
  description: >
    Public REST API for the EveryReview mobile clients (Android, iOS).
    Source of truth for the API contract — see ADR-0006 (REST) and
    ADR-0007 (versioning). Both mobile clients generate their network
    layer from this file.
servers:
  - url: https://api.everyreview.app/v1
    description: Production
  - url: https://api.staging.everyreview.app/v1
    description: Staging

tags:
  - name: auth
  - name: products
  - name: reviews

paths:
  /auth/anonymous:
    post:
      tags: [auth]
      summary: Create (or resume) an anonymous account for this device.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                device_id:
                  type: string
                  description: Client-generated stable device identifier.
              required: [device_id]
      responses:
        "200":
          description: Anonymous session issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"

  /auth/email/register:
    post:
      tags: [auth]
      summary: Link an email/password credential to the current account, or create a new account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
              required: [email, password]
      responses:
        "200":
          description: Credential linked, tokens issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
        "409":
          description: Email already in use.

  /auth/email/login:
    post:
      tags: [auth]
      summary: Log in with email/password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string }
              required: [email, password]
      responses:
        "200":
          description: Tokens issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"
        "401":
          description: Invalid credentials.

  /auth/refresh:
    post:
      tags: [auth]
      summary: Exchange a refresh token for a new access token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token: { type: string }
              required: [refresh_token]
      responses:
        "200":
          description: New tokens issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokens"

  /products/{barcode}:
    get:
      tags: [products]
      summary: Look up a product by barcode. Creates a placeholder if unknown.
      parameters:
        - name: barcode
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product found or placeholder created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductWithSummary"

  /products/{productId}/reviews:
    get:
      tags: [reviews]
      summary: List reviews for a product, newest first.
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: Paginated review list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewPage"
    post:
      tags: [reviews]
      summary: Submit a review for a product. Requires authentication (anonymous accounts allowed).
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewCreate"
      responses:
        "201":
          description: Review created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "429":
          description: Rate limited.

  /reviews/{reviewId}:
    delete:
      tags: [reviews]
      summary: Soft-delete a review (author only).
      security:
        - bearerAuth: []
      parameters:
        - name: reviewId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted.
        "403":
          description: Not the review's author.

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AuthTokens:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        access_token: { type: string }
        refresh_token: { type: string }
        expires_in: { type: integer, description: Access token TTL in seconds. }

    Product:
      type: object
      properties:
        id: { type: string, format: uuid }
        barcode: { type: string }
        name: { type: string, nullable: true }
        brand: { type: string, nullable: true }
        image_url: { type: string, nullable: true, description: CDN URL, not an S3 key. }
        is_placeholder: { type: boolean }

    RatingSummary:
      type: object
      properties:
        review_count: { type: integer }
        average_rating: { type: number, format: float, nullable: true }

    ProductWithSummary:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            rating_summary:
              $ref: "#/components/schemas/RatingSummary"

    Review:
      type: object
      properties:
        id: { type: string, format: uuid }
        product_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        body: { type: string }
        rating: { type: integer, nullable: true, minimum: 1, maximum: 5 }
        created_at: { type: string, format: date-time }

    ReviewCreate:
      type: object
      properties:
        body: { type: string, minLength: 1, maxLength: 5000 }
        rating: { type: integer, nullable: true, minimum: 1, maximum: 5 }
      required: [body]

    ReviewPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        next_cursor: { type: string, nullable: true }